minor changes
improved html composition. Added a background to the main page and its source in the footer. Minor improvements to the arrangement of elements on pages. Fixed a bug with adding an empty or text string to a numeric field (income/expense). added and updated about fields on the main page.